aurexis/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
httpx==0.27.0
aiohttp==3.9.5
pydantic==2.7.1
python-dotenv==1.0.1

# LLM providers
openai==1.30.1
anthropic==0.26.0
mistralai==0.4.2
groq==0.9.0

# Local LLM
llama-cpp-python==0.2.77

# Security
cryptography==42.0.7
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database / Storage
tinydb==4.8.0
aiosqlite==0.20.0

# Connectors
python-telegram-bot==21.3
discord.py==2.3.2
requests==2.32.2

# WebUI
jinja2==3.1.4
python-multipart==0.0.9
aiofiles==23.2.1

# Utilities
rich==13.7.1
typer==0.12.3
pyyaml==6.0.1
structlog==24.2.0
